remove all the remaining old c-style casts
git-svn-id: https://alcugs.almlys.org/svn/alcugs/trunk@2521 6ad0db0f-5bf4-0310-8fb6-eb40c444d375

+++ src/servers/vaultserver.cpp
@@ -189,7 +189,7 @@ namespace alc {
 			case NetMsgVaultTask:
 			{
 				bool isTask = (msg->cmd == NetMsgVaultTask);
-				tvMessage parsedMsg(isTask, /* 0 = non-TPOTS */(Byte)0);
+				tvMessage parsedMsg(isTask, /* 0 = non-TPOTS */0);
 				
 				// get the data out of the packet
 				tmVault vaultMsg(u);
